@@ -1,15 +1,17 @@
 module github.com/git-plm/gitplm
 
-go 1.19
+go 1.22
+
+toolchain go1.22.1
 
 require (
-	github.com/gocarina/gocsv v0.0.0-20211020200912-82fc2684cc48
+	github.com/gocarina/gocsv v0.0.0-20240520201108-78e41c74b4b1
+	github.com/otiai10/copy v1.9.0
 	github.com/samber/lo v1.33.0
 	gopkg.in/yaml.v2 v2.4.0
 )
 
 require (
-	github.com/otiai10/copy v1.9.0 // indirect
 	golang.org/x/exp v0.0.0-20220303212507-bbda1eaf7a17 // indirect
 	golang.org/x/sys v0.0.0-20220715151400-c0bba94af5f8 // indirect
 )
