@@ -17,24 +17,24 @@ remove:
  - ref: D13
  - ref: R11
 add:
- - cmpName: "screw #4,2"
+ - cmpName: "screw #4 2"
    ref: S3
    ipn: SCR-002-0002
 `
 
 var bomIn = `
-Ref;Qnty;Value;Cmp name;Footprint;Description;Vendor;IPN;Datasheet
-TP4, TP5;2;;Test point 2;;;;;
-R1, R2, ;2;;100K_100mw;;;;RES-006-0232;
-D1, D2, D13, D14;4;;diode;;;;DIO-023-0023;
-"R11, ";"1";"2010_500mW_1%_3000V_10M";"2010_500mW_1%_3000V_10M";"Resistor_SMD:R_2010_5025Metric";"";"";"RES-008-1005";"https://www.bourns.com/docs/Product-Datasheets/CHV.pdf"
+Ref,Qty,Value,Cmp name,Footprint,Description,Vendor,IPN,Datasheet
+TP4 TP5,2,,Test point 2,,,,,
+R1 R2,2,,100K_100mw,,,,RES-006-0232,
+D1 D2 D13 D14,4,,diode,,,,DIO-023-0023,
+"R11","1","2010_500mW_1%_3000V_10M","2010_500mW_1%_3000V_10M","Resistor_SMD:R_2010_5025Metric","","","RES-008-1005","https://www.bourns.com/docs/Product-Datasheets/CHV.pdf"
 `
 
 var bomExp = `
-Ref;Qnty;Value;Cmp name;Footprint;Description;Vendor;IPN;Datasheet
-D1, D2, D14;3;;diode;;;;DIO-023-0023;
-R1, R2;2;;100K_100mw;;;;RES-006-0232;
-S3;1;;screw #4,2;;;;SCR-002-0002;
+Ref,Qty,Value,Cmp name,Footprint,Description,Vendor,IPN,Datasheet
+D1 D2 D14,3,,diode,,,,DIO-023-0023,
+R1 R2,2,,100K_100mw,,,,RES-006-0232,
+S3,1,,screw #4 2,,,,SCR-002-0002,
 `
 
 func TestRelScript(t *testing.T) {
@@ -64,8 +64,8 @@ func TestRelScript(t *testing.T) {
 	}
 
 	if reflect.DeepEqual(bExp, bModified) != true {
-		t.Error("bExp not the same as bModified")
 		fmt.Printf("bExp: %v", bExp)
 		fmt.Printf("bModified: %v", bModified)
+		t.Error("bExp not the same as bModified")
 	}
 }
